server/requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
Django==3.1.3
django-rest-framework==0.1.0
djangorestframework==3.12.2
freeze==3.0
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.6.0
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
urllib3==1.26.2
wrapt==1.12.1
